Fix LSP errors (includes)
