+++ src/ECS/SubMesh.cpp
@@ -1,4 +1,5 @@
 #include <dz/ECS/SubMesh.hpp>
+#include <dz/GlobalUID.hpp>
 
 dz::ecs::SubMesh::SubMeshReflectable::SubMeshReflectable(const std::function<SubMesh*()>& get_submesh_function):
     get_submesh_function(get_submesh_function),
@@ -19,10 +20,7 @@ void* dz::ecs::SubMesh::SubMeshReflectable::GetVoidPropertyByIndex(int prop_inde
     assert(mesh_ptr);
     auto& mesh = *mesh_ptr;
     switch (prop_index) {
-    case 0: return &mesh.parent_index;
-    case 1: return &mesh.parent_cid;
-    case 2: return &mesh.mesh_index;
-    case 3: return &mesh.material_index;
+    case 0: return &mesh.material_index;
     default: return nullptr;
     }
 }
