+++ src/Loaders/STB_Image_Loader.cpp
@@ -1,19 +1,19 @@
 #include <dz/Loaders/STB_Image_Loader.hpp>
-
+#include <dz/Image.hpp>
 #define STB_IMAGE_IMPLEMENTATION
 #define STB_IMAGE_STATIC
 #include <stb_image.h>
 
-Image* STB_Image_load_rgba(const uint8_t* ptr, int width, int height) {
-    ImageCreateInfo info{
+dz::Image* STB_Image_load_rgba(const uint8_t* ptr, int width, int height) {
+    dz::ImageCreateInfo info{
         .width = (uint32_t)width,
         .height = (uint32_t)height,
         .data = (void*)ptr
     };
-    return image_create(info);
+    return dz::image_create(info);
 }
 
-Image* STB_Image_load_path(const std::filesystem::path& path) {
+dz::Image* STB_Image_load_path(const std::filesystem::path& path) {
 	int nrChannels;
     int width = 0, height = 0;
     std::string path_string = path.string();
@@ -24,7 +24,7 @@ Image* STB_Image_load_path(const std::filesystem::path& path) {
     return STB_Image_load_rgba(imageData, width, height);
 }
 
-Image* STB_Image_load_bytes(const std::shared_ptr<char>& bytes, size_t bytes_length) {
+dz::Image* STB_Image_load_bytes(const std::shared_ptr<char>& bytes, size_t bytes_length) {
 	int nrChannels;
     int width = 0, height = 0;
 	uint8_t *imageData = stbi_load_from_memory(
@@ -35,7 +35,7 @@ Image* STB_Image_load_bytes(const std::shared_ptr<char>& bytes, size_t bytes_len
     return STB_Image_load_rgba(imageData, width, height);
 }
 
-Image* dz::loaders::STB_Image_Loader::Load(const dz::loaders::STB_Image_Info& info) {
+dz::Image* dz::loaders::STB_Image_Loader::Load(const dz::loaders::STB_Image_Info& info) {
     if (!info.path.empty())
         return STB_Image_load_path(info.path);
     if (info.bytes && info.bytes_length)
