+++ src/path.cpp
@@ -1,3 +1,5 @@
+
+
 namespace dz {
     std::filesystem::path getUserDirectoryPath() { return std::filesystem::path(getenv("HOME")); }
 #if !defined(MACOS) && !defined(IOS)
