+++ tests/ECS.cpp
@@ -613,196 +613,212 @@ int main() {
                 ImGui::TextDisabled("GID: 0x%08X", SelectedReflectableID);
                 ImGui::Spacing();
 
-                auto& reflectables = reflectable_group.GetReflectables();
+                std::function<void(ReflectableGroup&)> render_reflectables;
+                render_reflectables = [&](auto& what_reflectable_group) {
+                    auto& reflectables = what_reflectable_group.GetReflectables();
 
-                auto reflect_begin = reflectables.begin();
-                auto reflect_it = reflect_begin;
-                auto reflect_end = reflectables.end();
-                size_t reflect_dist = 0;
+                    auto reflect_begin = reflectables.begin();
+                    auto reflect_it = reflect_begin;
+                    auto reflect_end = reflectables.end();
+                    size_t reflect_dist = 0;
 
-                auto update_iterators = [&]() {
-                    reflect_begin = reflectables.begin();
-                    reflect_it = reflect_begin + reflect_dist;
-                    reflect_end = reflectables.end();
+                    auto update_iterators = [&]() {
+                        reflect_begin = reflectables.begin();
+                        reflect_it = reflect_begin + reflect_dist;
+                        reflect_end = reflectables.end();
+                    };
+
+                    for (; reflect_it != reflect_end; reflect_it++) {
+                        reflect_dist = std::distance(reflect_begin, reflect_it);
+                        auto& reflectable = **reflect_it;
+                        ImGui::PushID(&reflectable);
+                        const auto& reflectable_name = reflectable.GetName();
+
+                        if (ImGui::CollapsingHeader(reflectable_name.c_str(), ImGuiTreeNodeFlags_DefaultOpen))
+                        {
+                            bool popped_style_var = false;
+                            ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, ImVec2(6, 4));
+                            ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, ImVec2(12, 8));
+
+                            auto reflectable_type_hint = reflectable.GetTypeHint();
+
+                            switch (reflectable_type_hint) {
+                            case ReflectableTypeHint::VEC2:
+                            {
+                                auto data_ptr = reflectable.GetVoidPropertyByIndex(0);
+                                if (ImGui::DragFloat2("##vec2", static_cast<float*>(data_ptr), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
+                                    reflectable.NotifyChange(0);
+                                    update_iterators();
+                                }
+                                break;
+                            }
+                            case ReflectableTypeHint::VEC3:
+                            {
+                                auto data_ptr = reflectable.GetVoidPropertyByIndex(0);
+                                if (ImGui::DragFloat3("##vec3", static_cast<float*>(data_ptr), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
+                                    reflectable.NotifyChange(0);
+                                    update_iterators();
+                                }
+                                break;
+                            }
+                            case ReflectableTypeHint::VEC3_RGB:
+                            {
+                                auto data_ptr = reflectable.GetVoidPropertyByIndex(0);
+                                if (ImGui::ColorEdit3("##rgb", static_cast<float*>(data_ptr), ImGuiColorEditFlags_Float)) {
+                                    reflectable.NotifyChange(0);
+                                    update_iterators();
+                                }
+                                break;
+                            }
+                            case ReflectableTypeHint::VEC4:
+                            {
+                                auto data_ptr = reflectable.GetVoidPropertyByIndex(0);
+                                if (ImGui::DragFloat4("##vec4", static_cast<float*>(data_ptr), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
+                                    reflectable.NotifyChange(0);
+                                    update_iterators();
+                                }
+                                break;
+                            }
+                            case ReflectableTypeHint::VEC4_RGBA:
+                            {
+                                auto data_ptr = reflectable.GetVoidPropertyByIndex(0);
+                                if (ImGui::ColorEdit4("##rgba", static_cast<float*>(data_ptr), ImGuiColorEditFlags_Float)) {
+                                    reflectable.NotifyChange(0);
+                                    update_iterators();
+                                }
+                                break;
+                            }
+                            default:
+                            case ReflectableTypeHint::STRUCT: {
+                                const auto& reflectable_typeinfos = reflectable.GetPropertyTypeinfos();
+                                const auto& reflectable_prop_names = reflectable.GetPropertyNames();
+                                const auto& disabled_properties = reflectable.GetDisabledProperties();
+                                size_t index = 0;
+                                for (auto& prop_name : reflectable_prop_names)
+                                {
+                                    ImGui::PushID(prop_name.c_str());
+                                    auto prop_index = reflectable.GetPropertyIndexByName(prop_name);
+                                    auto type_info = reflectable_typeinfos[prop_index];
+                                    bool is_disabled = (prop_index < disabled_properties.size()) ? disabled_properties[prop_index] : false;
+
+                                    ImGui::BeginDisabled(is_disabled);
+                                    ImGui::Text("%s", prop_name.c_str());
+                                    ImGui::SameLine();
+
+                                    if (*type_info == typeid(float))
+                                    {
+                                        auto& value = reflectable.GetPropertyByIndex<float>(prop_index);
+                                        ImGui::PushID(prop_index);
+                                        if (ImGui::InputFloat("##input", &value, 0.1f, 1.0f, "%.3f")) {
+                                            reflectable.NotifyChange(prop_index);
+                                            update_iterators();
+                                        }
+                                        ImGui::PopID();
+                                    }
+                                    else if (*type_info == typeid(int)) {
+                                        auto& value = reflectable.GetPropertyByIndex<int>(prop_index);
+                                        ImGui::PushID(prop_index);
+                                        if (ImGui::InputInt("##input", &value)) {
+                                            reflectable.NotifyChange(prop_index);
+                                            update_iterators();
+                                        }
+                                        ImGui::PopID();
+                                    }
+                                    else if (*type_info == typeid(Light::LightType)) {
+                                        auto& value = reflectable.GetPropertyByIndex<Light::LightType>(prop_index);
+                                        static const char* projection_types[] = { "Directional", "Spot", "Point" };
+                                        int current_index = static_cast<int>(value);
+                                        if (ImGui::Combo("##lightType", &current_index, projection_types, IM_ARRAYSIZE(projection_types))) {
+                                            value = static_cast<Light::LightType>(current_index);
+                                            reflectable.NotifyChange(prop_index);
+                                            update_iterators();
+                                        }
+                                    }
+                                    else if (*type_info == typeid(Camera::ProjectionType)) {
+                                        auto& value = reflectable.GetPropertyByIndex<Camera::ProjectionType>(prop_index);
+                                        static const char* projection_types[] = { "Perspective", "Orthographic" };
+                                        int current_index = static_cast<int>(value);
+                                        if (ImGui::Combo("##proj", &current_index, projection_types, IM_ARRAYSIZE(projection_types))) {
+                                            value = static_cast<Camera::ProjectionType>(current_index);
+                                            reflectable.NotifyChange(prop_index);
+                                            update_iterators();
+                                        }
+                                    }
+                                    else if (*type_info == typeid(std::string)) {
+                                        auto& value = reflectable.GetPropertyByIndex<std::string>(prop_index);
+                                        if (value.capacity() < 128) value.reserve(128);
+                                        ImGui::PushID(prop_index);
+                                        if (ImGui::InputText("##s", value.data(), value.capacity() + 1)) {
+                                            reflectable.NotifyChange(prop_index);
+                                            update_iterators();
+                                        }
+                                        ImGui::PopID();
+                                    }
+                                    else if (*type_info == typeid(vec<float, 2>)) {
+                                        auto& value = reflectable.GetPropertyByIndex<vec<float, 2>>(prop_index);
+                                        if (ImGui::DragFloat2("##vec2", static_cast<float*>(&value[0]), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
+                                            reflectable.NotifyChange(prop_index);
+                                            update_iterators();
+                                        }
+                                    }
+                                    else if (*type_info == typeid(vec<float, 3>)) {
+                                        auto& value = reflectable.GetPropertyByIndex<vec<float, 3>>(prop_index);
+                                        if (ImGui::DragFloat3("##vec3", static_cast<float*>(&value[0]), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
+                                            reflectable.NotifyChange(prop_index);
+                                            update_iterators();
+                                        }
+                                    }
+                                    else if (*type_info == typeid(vec<float, 4>)) {
+                                        auto& value = reflectable.GetPropertyByIndex<vec<float, 4>>(prop_index);
+                                        if (ImGui::DragFloat4("##vec4", static_cast<float*>(&value[0]), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
+                                            reflectable.NotifyChange(prop_index);
+                                            update_iterators();
+                                        }
+                                    }
+                                    else if (*type_info == typeid(color_vec<float, 3>)) {
+                                        auto& value = reflectable.GetPropertyByIndex<color_vec<float, 3>>(prop_index);
+                                        if (ImGui::ColorEdit3("##rgb", static_cast<float*>(&value[0]), ImGuiColorEditFlags_Float)) {
+                                            reflectable.NotifyChange(0);
+                                            update_iterators();
+                                        }
+                                    }
+                                    else if (*type_info == typeid(color_vec<float, 4>)) {
+                                        auto& value = reflectable.GetPropertyByIndex<color_vec<float, 4>>(prop_index);
+                                        if (ImGui::ColorEdit4("##rgba", static_cast<float*>(&value[0]), ImGuiColorEditFlags_Float)) {
+                                            reflectable.NotifyChange(0);
+                                            update_iterators();
+                                        }
+                                    }
+                                    else if (*type_info == typeid(MaterialIndexReflectable)) {
+                                        auto& value = reflectable.GetPropertyByIndex<MaterialIndexReflectable>(prop_index);
+                                        auto& material_group = ecs.GetGroupByIndex<Material, Material::ReflectableGroup>(value.material_index);
+                                        ImGui::EndDisabled();
+                                        ImGui::PopID();
+                                        ImGui::PopStyleVar(2);
+                                        popped_style_var = true;
+                                        render_reflectables(material_group);
+                                        break;
+                                    }
+                                    else {
+                                        ImGui::TextDisabled("<Unsupported type>");
+                                    }
+
+                                    ImGui::EndDisabled();
+                                    ImGui::PopID();
+                                }
+                                break;
+                            }
+                            }
+
+                            if (!popped_style_var)
+                                ImGui::PopStyleVar(2);
+                        }
+                    
+                        ImGui::PopID();
+                    }
                 };
 
-                for (; reflect_it != reflect_end; reflect_it++) {
-                    reflect_dist = std::distance(reflect_begin, reflect_it);
-                    auto& reflectable = **reflect_it;
-                    ImGui::PushID(&reflectable);
-                    const auto& reflectable_name = reflectable.GetName();
-
-                    if (ImGui::CollapsingHeader(reflectable_name.c_str(), ImGuiTreeNodeFlags_DefaultOpen))
-                    {
-                        ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, ImVec2(6, 4));
-                        ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, ImVec2(12, 8));
-
-                        auto reflectable_type_hint = reflectable.GetTypeHint();
-
-                        switch (reflectable_type_hint) {
-                        case ReflectableTypeHint::VEC2:
-                        {
-                            auto data_ptr = reflectable.GetVoidPropertyByIndex(0);
-                            if (ImGui::DragFloat2("##vec2", static_cast<float*>(data_ptr), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
-                                reflectable.NotifyChange(0);
-                                update_iterators();
-                            }
-                            break;
-                        }
-                        case ReflectableTypeHint::VEC3:
-                        {
-                            auto data_ptr = reflectable.GetVoidPropertyByIndex(0);
-                            if (ImGui::DragFloat3("##vec3", static_cast<float*>(data_ptr), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
-                                reflectable.NotifyChange(0);
-                                update_iterators();
-                            }
-                            break;
-                        }
-                        case ReflectableTypeHint::VEC3_RGB:
-                        {
-                            auto data_ptr = reflectable.GetVoidPropertyByIndex(0);
-                            if (ImGui::ColorEdit3("##rgb", static_cast<float*>(data_ptr), ImGuiColorEditFlags_Float)) {
-                                reflectable.NotifyChange(0);
-                                update_iterators();
-                            }
-                            break;
-                        }
-                        case ReflectableTypeHint::VEC4:
-                        {
-                            auto data_ptr = reflectable.GetVoidPropertyByIndex(0);
-                            if (ImGui::DragFloat4("##vec4", static_cast<float*>(data_ptr), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
-                                reflectable.NotifyChange(0);
-                                update_iterators();
-                            }
-                            break;
-                        }
-                        case ReflectableTypeHint::VEC4_RGBA:
-                        {
-                            auto data_ptr = reflectable.GetVoidPropertyByIndex(0);
-                            if (ImGui::ColorEdit4("##rgba", static_cast<float*>(data_ptr), ImGuiColorEditFlags_Float)) {
-                                reflectable.NotifyChange(0);
-                                update_iterators();
-                            }
-                            break;
-                        }
-                        default:
-                        case ReflectableTypeHint::STRUCT: {
-                            const auto& reflectable_typeinfos = reflectable.GetPropertyTypeinfos();
-                            const auto& reflectable_prop_names = reflectable.GetPropertyNames();
-                            const auto& disabled_properties = reflectable.GetDisabledProperties();
-                            size_t index = 0;
-                            for (auto& prop_name : reflectable_prop_names)
-                            {
-                                ImGui::PushID(prop_name.c_str());
-                                auto prop_index = reflectable.GetPropertyIndexByName(prop_name);
-                                auto type_info = reflectable_typeinfos[prop_index];
-                                bool is_disabled = (prop_index < disabled_properties.size()) ? disabled_properties[prop_index] : false;
-
-                                ImGui::BeginDisabled(is_disabled);
-                                ImGui::Text("%s", prop_name.c_str());
-                                ImGui::SameLine();
-
-                                if (*type_info == typeid(float))
-                                {
-                                    auto& value = reflectable.GetPropertyByIndex<float>(prop_index);
-                                    ImGui::PushID(prop_index);
-                                    if (ImGui::InputFloat("##input", &value, 0.1f, 1.0f, "%.3f")) {
-                                        reflectable.NotifyChange(prop_index);
-                                        update_iterators();
-                                    }
-                                    ImGui::PopID();
-                                }
-                                else if (*type_info == typeid(int)) {
-                                    auto& value = reflectable.GetPropertyByIndex<int>(prop_index);
-                                    ImGui::PushID(prop_index);
-                                    if (ImGui::InputInt("##input", &value)) {
-                                        reflectable.NotifyChange(prop_index);
-                                        update_iterators();
-                                    }
-                                    ImGui::PopID();
-                                }
-                                else if (*type_info == typeid(Light::LightType)) {
-                                    auto& value = reflectable.GetPropertyByIndex<Light::LightType>(prop_index);
-                                    static const char* projection_types[] = { "Directional", "Spot", "Point" };
-                                    int current_index = static_cast<int>(value);
-                                    if (ImGui::Combo("##lightType", &current_index, projection_types, IM_ARRAYSIZE(projection_types))) {
-                                        value = static_cast<Light::LightType>(current_index);
-                                        reflectable.NotifyChange(prop_index);
-                                        update_iterators();
-                                    }
-                                }
-                                else if (*type_info == typeid(Camera::ProjectionType)) {
-                                    auto& value = reflectable.GetPropertyByIndex<Camera::ProjectionType>(prop_index);
-                                    static const char* projection_types[] = { "Perspective", "Orthographic" };
-                                    int current_index = static_cast<int>(value);
-                                    if (ImGui::Combo("##proj", &current_index, projection_types, IM_ARRAYSIZE(projection_types))) {
-                                        value = static_cast<Camera::ProjectionType>(current_index);
-                                        reflectable.NotifyChange(prop_index);
-                                        update_iterators();
-                                    }
-                                }
-                                else if (*type_info == typeid(std::string)) {
-                                    auto& value = reflectable.GetPropertyByIndex<std::string>(prop_index);
-                                    if (value.capacity() < 128) value.reserve(128);
-                                    ImGui::PushID(prop_index);
-                                    if (ImGui::InputText("##s", value.data(), value.capacity() + 1)) {
-                                        reflectable.NotifyChange(prop_index);
-                                        update_iterators();
-                                    }
-                                    ImGui::PopID();
-                                }
-                                else if (*type_info == typeid(vec<float, 2>)) {
-                                    auto& value = reflectable.GetPropertyByIndex<vec<float, 2>>(prop_index);
-                                    if (ImGui::DragFloat2("##vec2", static_cast<float*>(&value[0]), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
-                                        reflectable.NotifyChange(prop_index);
-                                        update_iterators();
-                                    }
-                                }
-                                else if (*type_info == typeid(vec<float, 3>)) {
-                                    auto& value = reflectable.GetPropertyByIndex<vec<float, 3>>(prop_index);
-                                    if (ImGui::DragFloat3("##vec3", static_cast<float*>(&value[0]), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
-                                        reflectable.NotifyChange(prop_index);
-                                        update_iterators();
-                                    }
-                                }
-                                else if (*type_info == typeid(vec<float, 4>)) {
-                                    auto& value = reflectable.GetPropertyByIndex<vec<float, 4>>(prop_index);
-                                    if (ImGui::DragFloat4("##vec4", static_cast<float*>(&value[0]), 0.01f, -1000.0f, 1000.0f, "%.3f", ImGuiSliderFlags_AlwaysClamp)) {
-                                        reflectable.NotifyChange(prop_index);
-                                        update_iterators();
-                                    }
-                                }
-                                else if (*type_info == typeid(color_vec<float, 3>)) {
-                                    auto& value = reflectable.GetPropertyByIndex<color_vec<float, 3>>(prop_index);
-                                    if (ImGui::ColorEdit3("##rgb", static_cast<float*>(&value[0]), ImGuiColorEditFlags_Float)) {
-                                        reflectable.NotifyChange(0);
-                                        update_iterators();
-                                    }
-                                }
-                                else if (*type_info == typeid(color_vec<float, 4>)) {
-                                    auto& value = reflectable.GetPropertyByIndex<color_vec<float, 4>>(prop_index);
-                                    if (ImGui::ColorEdit4("##rgba", static_cast<float*>(&value[0]), ImGuiColorEditFlags_Float)) {
-                                        reflectable.NotifyChange(0);
-                                        update_iterators();
-                                    }
-                                }
-                                else {
-                                    ImGui::TextDisabled("<Unsupported type>");
-                                }
-
-                                ImGui::EndDisabled();
-                                ImGui::PopID();
-                            }
-                            break;
-                        }
-                        }
-
-                        ImGui::PopStyleVar(2);
-                    }
-                
-                    ImGui::PopID();
-                }
-
+                render_reflectables(reflectable_group);
                 ImGui::PopID();
             }
